fix: pr etst

+++ src/main/java/com/app/pingpong/domain/image/S3Uploader.java
@@ -44,7 +44,8 @@ public class S3Uploader {
                 throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "파일 업로드에 실패했습니다.");
             }
 
-            fileNameList.add(fileName);
+            fileNameList.add(amazonS3.getUrl(bucket, fileName).toString());
+
         });
         return fileNameList;
     }
@@ -57,7 +58,13 @@ public class S3Uploader {
     }
 
     public void deleteFile(String fileName) {
-        amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
+        if (!fileName.isEmpty() || !fileName.equals("")) {
+            fileName = fileName.substring(fileName.lastIndexOf("/") + 1);
+
+            if (amazonS3.doesObjectExist(bucket, fileName)) {
+                amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
+            }
+        }
     }
 
     public String getFilePath(String path) {
